requirements.txt
supabase
textblob
python-dotenv
requests
rich